+++ CRM/Civirules/Upgrader.php
@@ -847,5 +847,11 @@ class CRM_Civirules_Upgrader extends CRM_Civirules_Upgrader_Base {
     return TRUE;
   }
 
+  public function upgrade_2074() {
+    $this->ctx->log->info('Applying update 2074 - Add actions add set custom field with data from another custom field');
+    CRM_Civirules_Utils_Upgrader::insertActionsFromJson($this->extensionDir . DIRECTORY_SEPARATOR . 'sql/actions.json');
+    return TRUE;
+  }
+
 }